+++ projects/rccl/plugins/profiler/accl/accl_shim.h
@@ -1,19 +1,14 @@
-/*
- * Copyright (c) 2016-2024, NVIDIA CORPORATION. All rights reserved.
- * Copyright (c) 2026 Advanced Micro Devices, Inc. All rights reserved.
+/*************************************************************************
+ * SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES.
+ *                         All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ * Modifications Copyright (c) 2026 Advanced Micro Devices, Inc. All rights reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * See LICENSE.txt for more license information
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+ * Minimal profiler-API types for the accl-profiler plugin only. Values must stay in sync with
+ * src/include/plugin/nccl_profiler.h and src/include/nccl_common.h.
+ *************************************************************************/
 
 #ifndef ACCL_SHIM_H_
 #define ACCL_SHIM_H_